Haunt/scripts/github-integration/requirements.txt
# GitHub Integration Dependencies

# Web framework for webhook receiver
flask>=3.0.0

# HTTP requests for GitHub API
requests>=2.31.0

# YAML configuration parsing
PyYAML>=6.0

# Testing dependencies (optional, for development)
pytest>=7.4.0
pytest-mock>=3.12.0
